refactor: update context file names
